fix: implement the metadata properly
